pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discordboy"
version = "0.1.0"
description = "A Discord bot that runs a Game Boy emulator and allows users to play games collectively"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Discord GameBoy Bot"}
]
dependencies = [
    "discord.py>=2.4.0",
    "pyboy>=2.4.1",
    "Pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.0",
    "pre-commit>=3.6.0",
    "mypy>=1.8.0",
    "bandit[toml]>=1.7.7",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
discordboy = "discordboy.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["discordboy*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --cov=discordboy --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["discordboy"]
omit = ["*/tests/*", "*/__pycache__/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
